rename "compositor"/"surface" to "canvaskit"/"html"; document conventions
