+++ lib/web_ui/lib/src/engine.dart
@@ -27,43 +27,70 @@ part 'engine/assets.dart';
 part 'engine/bitmap_canvas.dart';
 part 'engine/browser_detection.dart';
 part 'engine/browser_location.dart';
+part 'engine/canvaskit/canvas.dart';
+part 'engine/canvaskit/canvaskit_canvas.dart';
+part 'engine/canvaskit/canvaskit_api.dart';
+part 'engine/canvaskit/color_filter.dart';
+part 'engine/canvaskit/embedded_views.dart';
+part 'engine/canvaskit/fonts.dart';
+part 'engine/canvaskit/image.dart';
+part 'engine/canvaskit/image_filter.dart';
+part 'engine/canvaskit/initialization.dart';
+part 'engine/canvaskit/layer.dart';
+part 'engine/canvaskit/layer_scene_builder.dart';
+part 'engine/canvaskit/layer_tree.dart';
+part 'engine/canvaskit/mask_filter.dart';
+part 'engine/canvaskit/n_way_canvas.dart';
+part 'engine/canvaskit/path.dart';
+part 'engine/canvaskit/painting.dart';
+part 'engine/canvaskit/path_metrics.dart';
+part 'engine/canvaskit/picture.dart';
+part 'engine/canvaskit/picture_recorder.dart';
+part 'engine/canvaskit/platform_message.dart';
+part 'engine/canvaskit/raster_cache.dart';
+part 'engine/canvaskit/rasterizer.dart';
+part 'engine/canvaskit/shader.dart';
+part 'engine/canvaskit/skia_object_cache.dart';
+part 'engine/canvaskit/surface.dart';
+part 'engine/canvaskit/text.dart';
+part 'engine/canvaskit/util.dart';
+part 'engine/canvaskit/vertices.dart';
+part 'engine/canvaskit/viewport_metrics.dart';
 part 'engine/canvas_pool.dart';
 part 'engine/clipboard.dart';
 part 'engine/color_filter.dart';
-part 'engine/compositor/canvas.dart';
-part 'engine/compositor/canvas_kit_canvas.dart';
-part 'engine/compositor/canvaskit_api.dart';
-part 'engine/compositor/color_filter.dart';
-part 'engine/compositor/embedded_views.dart';
-part 'engine/compositor/fonts.dart';
-part 'engine/compositor/image.dart';
-part 'engine/compositor/image_filter.dart';
-part 'engine/compositor/initialization.dart';
-part 'engine/compositor/layer.dart';
-part 'engine/compositor/layer_scene_builder.dart';
-part 'engine/compositor/layer_tree.dart';
-part 'engine/compositor/mask_filter.dart';
-part 'engine/compositor/n_way_canvas.dart';
-part 'engine/compositor/path.dart';
-part 'engine/compositor/painting.dart';
-part 'engine/compositor/path_metrics.dart';
-part 'engine/compositor/picture.dart';
-part 'engine/compositor/picture_recorder.dart';
-part 'engine/compositor/platform_message.dart';
-part 'engine/compositor/raster_cache.dart';
-part 'engine/compositor/rasterizer.dart';
-part 'engine/compositor/shader.dart';
-part 'engine/compositor/skia_object_cache.dart';
-part 'engine/compositor/surface.dart';
-part 'engine/compositor/text.dart';
-part 'engine/compositor/util.dart';
-part 'engine/compositor/vertices.dart';
-part 'engine/compositor/viewport_metrics.dart';
 part 'engine/dom_canvas.dart';
 part 'engine/dom_renderer.dart';
 part 'engine/engine_canvas.dart';
 part 'engine/frame_reference.dart';
 part 'engine/history.dart';
+part 'engine/html/backdrop_filter.dart';
+part 'engine/html/canvas.dart';
+part 'engine/html/clip.dart';
+part 'engine/html/debug_canvas_reuse_overlay.dart';
+part 'engine/html/image_filter.dart';
+part 'engine/html/offset.dart';
+part 'engine/html/opacity.dart';
+part 'engine/html/painting.dart';
+part 'engine/html/path/conic.dart';
+part 'engine/html/path/cubic.dart';
+part 'engine/html/path/path.dart';
+part 'engine/html/path/path_metrics.dart';
+part 'engine/html/path/path_ref.dart';
+part 'engine/html/path/path_to_svg.dart';
+part 'engine/html/path/path_utils.dart';
+part 'engine/html/path/path_windings.dart';
+part 'engine/html/path/tangent.dart';
+part 'engine/html/picture.dart';
+part 'engine/html/platform_view.dart';
+part 'engine/html/recording_canvas.dart';
+part 'engine/html/render_vertices.dart';
+part 'engine/html/scene.dart';
+part 'engine/html/scene_builder.dart';
+part 'engine/html/shader.dart';
+part 'engine/html/surface.dart';
+part 'engine/html/surface_stats.dart';
+part 'engine/html/transform.dart';
 part 'engine/html_image_codec.dart';
 part 'engine/keyboard.dart';
 part 'engine/mouse_cursor.dart';
@@ -91,33 +118,6 @@ part 'engine/services/message_codec.dart';
 part 'engine/services/message_codecs.dart';
 part 'engine/services/serialization.dart';
 part 'engine/shadow.dart';
-part 'engine/surface/backdrop_filter.dart';
-part 'engine/surface/canvas.dart';
-part 'engine/surface/clip.dart';
-part 'engine/surface/debug_canvas_reuse_overlay.dart';
-part 'engine/surface/image_filter.dart';
-part 'engine/surface/offset.dart';
-part 'engine/surface/opacity.dart';
-part 'engine/surface/painting.dart';
-part 'engine/surface/path/conic.dart';
-part 'engine/surface/path/cubic.dart';
-part 'engine/surface/path/path.dart';
-part 'engine/surface/path/path_metrics.dart';
-part 'engine/surface/path/path_ref.dart';
-part 'engine/surface/path/path_to_svg.dart';
-part 'engine/surface/path/path_utils.dart';
-part 'engine/surface/path/path_windings.dart';
-part 'engine/surface/path/tangent.dart';
-part 'engine/surface/picture.dart';
-part 'engine/surface/platform_view.dart';
-part 'engine/surface/recording_canvas.dart';
-part 'engine/surface/render_vertices.dart';
-part 'engine/surface/scene.dart';
-part 'engine/surface/scene_builder.dart';
-part 'engine/surface/shader.dart';
-part 'engine/surface/surface.dart';
-part 'engine/surface/surface_stats.dart';
-part 'engine/surface/transform.dart';
 part 'engine/test_embedding.dart';
 part 'engine/text/font_collection.dart';
 part 'engine/text/line_break_properties.dart';
