+++ lib/web_ui/lib/src/engine/canvaskit/path_metrics.dart
